pyproject.toml
[tool.poetry]
name = "agro-verx"
version = "0.1.0"
description = ""
authors = ["Renan <renanhc96@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9"
python-decouple = "^3.8"
drf-yasg = "^1.21.7"
setuptools = "^72.1.0"
gunicorn = "^22.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
blue = "^0.9.1"
isort = "^5.13.2"
taskipy = "^1.13.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"
line_length = 79

[tool.taskipy.tasks]
lint = "blue . && isort ."
pre_test = "task lint"
test = "pytest -s -x --cov=agro_verx -vv"
post_test = "coverage html"
loaddata = "python manage.py loaddata core_state core_city producer_producer producer_farms producer_plantation_type producer_farm_plantation"
